Frontend clean up schema
