+++ frontend/src/visualization/pages/Graph.tsx
@@ -1,51 +1,19 @@
 import React, { useState } from 'react';
 import VisGraph from 'react-graph-vis';
-import { createStyles, makeStyles } from '@material-ui/core/styles';
 import { uuid } from 'uuidv4';
 import { map, tap } from 'rxjs/operators';
 import { combineLatest } from 'rxjs';
 import { Snackbar } from '@material-ui/core';
 import { Alert } from '@material-ui/lab';
-import useService from '../dependency-injection/useService';
-import { ContainerSize } from '../utils/useSize';
-import useObservable from '../utils/useObservable';
-import QueryResultStore from '../stores/QueryResultStore';
-import convertQueryResult from './shared-ops/convertQueryResult';
-import { createSelectionInfo, EntityStyleStore } from '../stores/colors';
-import SearchSelectionStore from '../stores/SearchSelectionStore';
-import { isEntitySelected } from '../stores/colors/EntityStyleProviderImpl';
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    graphContainer: {
-      zIndex: 1200,
-      position: 'relative',
-      flexGrow: 1,
-      overflowY: 'hidden',
-      overflowX: 'hidden',
-    },
-  })
-);
-
-/**
- * Builds the graph options passed to react-graph-vis.
- * @param width The width of the graph.
- * @param height The height of the graph.
- * @param layout Possible values: "hierarchical", undefined
- * @returns The react-graph-vis options.
- */
-function buildOptions(width: number, height: number, layout?: string) {
-  return {
-    layout: {
-      hierarchical: layout === 'hierarchical',
-    },
-    edges: {
-      color: '#000000',
-    },
-    width: `${width}px`,
-    height: `${height}px`,
-  };
-}
+import useService from '../../dependency-injection/useService';
+import { ContainerSize } from '../../utils/useSize';
+import useObservable from '../../utils/useObservable';
+import QueryResultStore from '../../stores/QueryResultStore';
+import convertQueryResult from '../shared-ops/convertQueryResult';
+import { createSelectionInfo, EntityStyleStore } from '../../stores/colors';
+import SearchSelectionStore from '../../stores/SearchSelectionStore';
+import { isEntitySelected } from '../../stores/colors/EntityStyleProviderImpl';
+import { buildOptions, useStylesVisualization } from './shared-options';
 
 type GraphProps = {
   layout?: string;
@@ -54,7 +22,7 @@ type GraphProps = {
 
 function Graph(props: GraphProps): JSX.Element {
   const { layout, containerSize } = props;
-  const classes = useStyles();
+  const classes = useStylesVisualization();
 
   // whether snackbar with hint that selected entity is not found is open or not
   const [noEntitiesFoundWarningOpen, setNoEntitiesFoundWarningOpen] =
