+++ frontend/src/visualization/pages/SchemaPage.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
-import { ContainerSize } from '../utils/useSize';
+import { ContainerSize } from '../../utils/useSize';
 import Schema from './Schema';
 import GraphPage from './GraphPage';
 
+/**
+ * Contains the schema visualization.
+ */
 function SchemaPage(): JSX.Element {
   const content = (containerSize: ContainerSize) => (
     <Schema containerSize={containerSize} />
