+++ frontend/src/visualization/pages/shared-options.ts
@@ -1,0 +1,34 @@
+import { createStyles, makeStyles } from '@material-ui/core/styles';
+
+export const useStylesVisualization = makeStyles(() =>
+  createStyles({
+    graphContainer: {
+      zIndex: 1200,
+      position: 'relative',
+      flexGrow: 1,
+      overflowY: 'hidden',
+      overflowX: 'hidden',
+    },
+  })
+);
+
+/**
+ * Builds the graph options passed to react-graph-vis.
+ * @param width The width of the graph.
+ * @param height The height of the graph.
+ * @param layout Possible values: "hierarchical", undefined
+ * @returns The react-graph-vis options.
+ */
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export function buildOptions(width: number, height: number, layout?: string) {
+  return {
+    layout: {
+      hierarchical: layout === 'hierarchical',
+    },
+    edges: {
+      color: '#000000',
+    },
+    width: `${width}px`,
+    height: `${height}px`,
+  };
+}
